Feat: Add formatter for GCS writter
